Started writing guide to creating a new trainer
