+++ doc/docs/reference/running_experiments.rst
@@ -25,7 +25,10 @@ all the experiment components, running the experiment, and saving the results. I
 designed to be as simple as possible, while still allowing for a wide range of
 experiments.
 
-See :doc:`../guides/running_experiments` for a more detailed guide on running experiments.
+See :doc:`../guides/running_experiments` for a more detailed guide on running
+experiments and :doc:`../guides/rexperiment-build-process` for an overview of how the
+experiment components are built by the :func:`run_experiment() <nip.run.run_experiment>`
+function.
 
 
 Example
