+++ nip/trainers/__init__.py
@@ -4,7 +4,7 @@ A trainer takes the components of a scenario and trains the agents.
 """
 
 from nip.parameters import HyperParameters
-from nip.factory import ScenarioInstance
+from nip.scenario_instance import ScenarioInstance
 from nip.experiment_settings import ExperimentSettings
 
 from .trainer_base import (
@@ -13,7 +13,7 @@ from .trainer_base import (
     IterationContext,
     attach_progress_bar,
 )
-from .rl_tensordict_base import ReinforcementLearningTrainer
+from .rl_tensordict_base import TensorDictRlTrainer
 from .vanilla_ppo import VanillaPpoTrainer
 from .solo_agent import SoloAgentTrainer
 from .spg import SpgTrainer
@@ -22,6 +22,22 @@ from .rl_pure_text_base import PureTextRlTrainer
 from .ei_pure_text import PureTextEiTrainer
 from .malt_pure_text import PureTextMaltTrainer
 from .registry import register_trainer, TRAINER_REGISTRY
+
+
+def get_trainer_class(hyper_params: HyperParameters) -> type[Trainer]:
+    """Get the trainer class from the hyperparameters.
+
+    Parameters
+    ----------
+    hyper_params : HyperParameters
+        The parameters of the experiment.
+
+    Returns
+    -------
+    trainer_class : type[Trainer]
+        The trainer class.
+    """
+    return TRAINER_REGISTRY[hyper_params.trainer]
 
 
 def build_trainer(
@@ -40,6 +56,4 @@ def build_trainer(
     settings : ExperimentSettings
         The instance-specific settings of the experiment, like device, logging, etc.
     """
-    return TRAINER_REGISTRY[hyper_params.trainer](
-        hyper_params, scenario_instance, settings
-    )
+    return get_trainer_class(hyper_params)(hyper_params, scenario_instance, settings)
