+++ requirements_dev.txt
@@ -13,8 +13,9 @@ sphinx==8.1.3
 sphinx-autobuild==2024.10.3
 sphinx-book-theme==1.1.3
 sphinx_tabs==3.4.7
-sphinxcontrib-bibtex==2.6.3
 sphinxcontrib-autoprogram==0.1.9
+sphinxcontrib-bibtex==2.6.3
+sphinxcontrib-mermaid==1.0.0
 sympy==1.12.1
 tensorboard==2.17.0
 torch_tb_profiler==0.4.3
